+++ test/mri/excludes/TestRDocOptions.rb
@@ -1,0 +1,2 @@
+exclude :test_parse_root, "slightly different path inspect string"
+exclude :test_parse_page_dir, "slightly different path inspect string"