+++ CHANGELOG.md
@@ -7,6 +7,10 @@ New features:
 * `Polyglot.export` can now be used with primitives, and will now convert
   strings to Java, and `.import` will convert them from Java.
 * Implemented `--encoding`, `--external-encoding`, `--internal-encoding`.
+* `rb_object_tainted` and similar C functions have been implemented.
+* `rb_struct_define_under` has been implemented.
+* `RbConfig::CONFIG['sysconfdir']` has been implemented.
+* More of `Etc` has been implemented.
 
 Bug fixes: